fix(repos): fix related structs for CVDetails

+++ src/models/range_values.rs
@@ -4,6 +4,6 @@ use serde::{Serialize, Deserialize};
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
 
 pub struct RangeValues{
-    pub lower: i32,
-    pub upper: i32
+    pub lower: f64,
+    pub upper: f64
 }
